Graph area now draws labels for vertical lines.

+++ graphene/src/main/java/org/epics/graphene/Graph2DArea.java
@@ -58,7 +58,7 @@ public interface Graph2DArea {
     List<String> getXReferenceLabels();
     Font getXReferenceLabelFont();
     Color getXReferenceLabelColor();
-    double getXReferenceLabelMargin();
+    int getXReferenceLabelMargin();
     
     ListNumber getYReferences();
     ListNumber getYReferenceValues();
